Refactor DefaultParatextProjectSettings to its own file
