+++ tests/SIL.Machine.Tests/Corpora/MemoryParatextProjectFileHandler.cs
@@ -1,5 +1,4 @@
 using System.Text;
-using SIL.Scripture;
 
 namespace SIL.Machine.Corpora;
 
@@ -23,41 +22,4 @@ public class MemoryParatextProjectFileHandler(IDictionary<string, string>? files
             return null;
         return new MemoryStream(Encoding.UTF8.GetBytes(contents));
     }
-
-    public class DefaultParatextProjectSettings(
-        string id = "Id",
-        string name = "Test",
-        string fullName = "TestProject",
-        Encoding? encoding = null,
-        ScrVers? versification = null,
-        UsfmStylesheet? stylesheet = null,
-        string fileNamePrefix = "",
-        string fileNameForm = "41MAT",
-        string fileNameSuffix = "Test.SFM",
-        string biblicalTermsListType = "Project",
-        string biblicalTermsProjectName = "Test",
-        string biblicalTermsFileName = "ProjectBiblicalTerms.xml",
-        string languageCode = "en",
-        string translationType = "Standard",
-        string? parentGuid = null,
-        string? parentName = null
-    )
-        : ParatextProjectSettings(
-            id,
-            name,
-            fullName,
-            encoding ?? Encoding.UTF8,
-            versification ?? ScrVers.English,
-            stylesheet ?? new UsfmStylesheet("usfm.sty"),
-            fileNamePrefix,
-            fileNameForm,
-            fileNameSuffix,
-            biblicalTermsListType,
-            biblicalTermsProjectName,
-            biblicalTermsFileName,
-            languageCode,
-            translationType,
-            parentGuid,
-            parentName
-        );
 }
