+++ tests/SIL.Machine.Tests/Corpora/MemoryParatextProjectTermsParser.cs
@@ -3,5 +3,5 @@ namespace SIL.Machine.Corpora;
 public class MemoryParatextProjectTermsParser(IDictionary<string, string>? files, ParatextProjectSettings? settings)
     : ParatextProjectTermsParserBase(
         new MemoryParatextProjectFileHandler(files),
-        settings ?? new MemoryParatextProjectFileHandler.DefaultParatextProjectSettings()
+        settings ?? new DefaultParatextProjectSettings()
     ) { }
