+++ tests/SIL.Machine.Tests/Corpora/MemoryParatextProjectTextUpdater.cs
@@ -1,4 +1,10 @@
 namespace SIL.Machine.Corpora;
 
-public class MemoryParatextProjectTextUpdater(IDictionary<string, string>? files, ParatextProjectSettings settings)
-    : ParatextProjectTextUpdaterBase(new MemoryParatextProjectFileHandler(files), settings);
+public class MemoryParatextProjectTextUpdater(
+    IDictionary<string, string>? files = null,
+    ParatextProjectSettings? settings = null
+)
+    : ParatextProjectTextUpdaterBase(
+        new MemoryParatextProjectFileHandler(files),
+        settings ?? new DefaultParatextProjectSettings()
+    );
